Extract competenceAreas as a separate component
