+++ apps/season-editor/src/App.tsx
@@ -6,13 +6,13 @@ import "@fontsource/roboto/700.css";
 import { CssBaseline, ThemeProvider } from "@mui/material";
 
 import theme from "./theme";
-import { AppContainer } from "./components/app-container";
+import { SeasonEditor } from "./pages/season-editor";
 
 function App() {
   return (
     <ThemeProvider theme={theme}>
       <CssBaseline />
-      <AppContainer />
+      <SeasonEditor />
     </ThemeProvider>
   );
 }
