+++ apps/season-editor/src/pages/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./competence-areas";
